issue https://adempiere.atlassian.net/browse/ADEMPIERE-311
The method updateInventoryValue() is only defined in the abstract class AbstractCostingMethod.
It is not defined in the class AverageInvoiceCostingMethod nor in the class StandardCostingMethod.

This means, updateInventoryValue() is performed the same way for both costing methods.
As the method updateInventoryValue() in the abstract class AbstractCostingMethod calculates average for currentCostPrice, this happens also for Standard Costing.
This causes the error stated in issue 311 and it should not happen.

The error is avoided by copying the method updateInventoryValue() from the AbstractCostingMethod class to the AverageInvoiceCostingMethod class AND by taking off the average cost calculation in updateInventoryValue() of the AbstractCostingMethod class.

This way, Average invoice calculates average for currentCostPrice and Standard Costing not.

Thanks to Susanne for the great support in finding and solving.