Average Invoice: the actual stock is base for the recalculation of Landed Costs of Match Invoice instead of the purchased quantity. Before this fix the full amount of the landed cost invoice was added to the inventory and included in the the new average cost. Example: A landed cost invoice is generated and distributed, but 50% of the purchased quantity is already sold. If you add the full amount of the landed cost to the remaining stock, you inflate your inventory value and your costs of sale are too small.
Costengine Calculating Costs of products, produced with MProduction:
Before: the cost of the produced product is the sum of the MCostdetail entries of the Production lines.
Error: the produced product is the first entry of the production lines, so the cost of the produced product is null.
To avoid unbalanced error in DOC_Production, the precision for the accounting is already included here.

Process org.compiere.process.ProductionCreate:
In the new costengine the costing method is defined in MCosttype, not in MCostelement, as before.
To use this process for other costing Methods than Standard Costing a new parameter M_Costtype is included.