ADEMPIERE-384 #resolve #comment ZK is giving "message not found errors"
The AD_Message and clear text messages were getting concatenated and translated a second time in the Ask dialog.  This commit changes the Ask functions to follow the same pattern as the info, warn and error functions and makes use of the private constructMessage function to generate the message text.