#1926 ZK Exports migration XML files to different location than what is selected in the dialogs. https://github.com/adempiere/adempiere/issues/1926
Removed file name as a parameter from the Export Migration to XML process.  Changed ProcessInfo to include a field for interface type with getters and setters.  Set the interface type in the process dialogs in both SWING and ZK.  Use the interface type in the process to trigger the correct handling of the file download process.  In ZK, use reflexion to dymanically load the correct library.  In Swing, the library is available.

(cherry picked from commit 6935bfebddb6c7cce66e908d54b6324b7880eb33)