#431 Fixes the way totals are grouped in the report dataengine.
Groupings were done in the order of columns not the order of sorting/grouping.  This change adds a recursive method that checks the groupings in the correct order and adds the summary functions, if any.  Groups can be displayed in any column and the summaries will respect the sort order/groupings.