Fixed error with inconsistency of payment
## Step by Step
- Create a Payment and complete: Amount **USD 1500**
- Add payment to bank statement: Account balance **USD 1500**
- Reverse payment
- See the account balance: **USD 1500**
- See Payment Details (parameter Is Reconciled true) report: **USD
1500**


## The problem
After reverse it is created other payment as reversal (it is ok) the
problem is that the original payment and reversal is flagged as un
reconciled.

### What is the result?
- The Account balance is of **USD 1500**
- The Payment Details with flag Is Reconciled as true: **USD 0**

Both must match but is not really

## What is the expected behavior?
Both (Bank Balance and Payment details as reconciled) must be **USD
1500**, the reversed payment will should be flagged as not reconciled
instead and it can be available for reconciled with a new bank
statement.

## Other problem?
the allocate flag is not set after reversed for both payments
nevertheless, the payment allocation is created and completed

# The follows are some images of error: