Update IT_Query to be more robust in the case where there are multiple clients in the database.  The previous tests assumed the seed was being used for the test which may not always be correct. (+1 squashed commits)
Change the way the message is checked when testing for mandatory columns in models.  This may be database specific (Postgres). (+1 squashed commits)

Update the base build.xml to incorporate classes in org.eevolution.manufacturing module in the test class path.  Add a null check in the Doc.savePostedStatus method.  And trap the null document, if one is returned by getDoc in the SessionPoster. (+1 squashed commits)

The method getDateAcct uses an out-of-date list of models that have non-standard Date Accts.  Update FactAcctReset to use the new method Doc.getDateAcctColumnName rather than hard-coding the column names. (+1 squashed commits)

Update MPeriod to fix merge error.