Improve modification indication UX in Swing editor
Instead of indicating that data is modified only when focus is out of
editor, this commit try to mark modification immediately after text in
editor is changed.

With this change, user can type <F4> or click Save icon directly to save
data after text editing is finished, without having to move focus out of
editor.