Fix broken input method in single-row mode in Swing UI
In single-row mode, text input widgets and models behind them have
kind of two-way data bindings, which are established in GridController.

In order to mark data as modified as soon as possible. we currently
update model data more eagerly, as commited in #3592, which will cause
problem, i.e.:

1. user type characters in text widget
2. model got updated because of vetoable change
3. text widget got updated again because of property change

The third step not only adds an unneeded update operation, but also
cause problem if interfered with on-the-spot input methods, which will
make uncommited characters to be saved in the text fields.

This commit tries to fix this problem by making sure that text widget
only need to be updated if value is different from model.

Fixes #3613