Fixed error with run process without transaction
# The Problem
When a process is running without transaction using the
`ProcessBuilder`, this execution make a new transaction for it.

All is ok but when a workflow is processed the transaction is openend
and never ending.

# The solution

Just let the process handle the own transaction if the `ProcessBuilder`
is called without transaction

# Step for Step
A example for run

```Java
ProcessBuilder builder = ProcessBuilder.create(Env.getCtx())
			.process(111)	//	C_Invoice Process
			.withRecordId(318, 100)	//
			.withoutPrintPreview()
			.withWindowNo(0)
			.withTitle("Process Invoice")
			.withParameter("C_Invoice_ID", recordId)
			.withParameter("DocAction", "CO");
		//	Execute Process
builder.execute(null);
```